changed org_id and activation_id to username and passwod for standard registration option
added pool-id to subscription and refactored

+++ testing/reposync_server.sh
@@ -29,43 +29,47 @@ function install_deps() {
     run_cmd dnf -y install httpd dnf-utils createrepo
     run_cmd dnf -y install mod_ssl
     echo "Dependencies installed..."
+    echo " "
 }
 
 function install_epel() {
     run_cmd dnf -y install $EPEL_RPM_PUBLIC_URL
     echo "Epel-release installed..."
+    echo " "
+}
+
+function clean_subs() {
+            subscription-manager remove --all
+            subscription-manager unregister
+            subscription-manager clean
 }
 
 function subscription_prompts() {
     echo "Verifying RedHat Subscription..."
     subscription_status=`subscription-manager status | grep 'Overall Status:' | awk '{ print $3 }'`
-
     if [ "$subscription_status" != "Current" ]; then
         echo "-------"
         echo "Since you are running a RHEL controller outside the Dev Network and/or not using Labrepo, "
         echo "You will need to subscribe to RHEL repositories."
         echo "-------"
         echo "Select RedHat subscription method:"
+        echo " "
         echo "Standard: Requires Org + Activation Key"
-        echo "RedHat Developer Login: A RedHat Developer account is free signup here https://developers.redhat.com/"
-        echo "RedHat Developer License cannot be used in production environments"
-        select cr in "Standard" "RedHat Developer" ; do
+        echo "RedHat Login: Requires User + Password"
+
+        select cr in "Standard" "RedHat Login" ; do
                 case $cr in
                     Standard )
                     export RHEL_SUB_METHOD="standard";
                     break
                     ;;
-                    "RedHat Developer" )
-                    export RHEL_SUB_METHOD="developer";
+                    "RedHat Login" )
+                    export RHEL_SUB_METHOD="login";
                     break
                     ;;
                 esac
             done
-        while true; do
-            subscription-manager remove --all
-            subscription-manager unregister
-            subscription-manager clean
-
+            clean_subs
             if [ "$RHEL_SUB_METHOD" == "standard" ]; then
                 if [ -z "$RHEL_ORGANIZATION" ]; then
                     read -p 'Please enter your RHEL org number (EX: Its the --org flag for the subscription-manager command): ' orgnumber
@@ -77,23 +81,11 @@ function subscription_prompts() {
                     export RHEL_ACTIVATIONKEY=$activationkey
                 fi
                 subscription-manager register --activationkey=$RHEL_ACTIVATIONKEY --org=$RHEL_ORGANIZATION --force
-            elif [ "$RHEL_SUB_METHOD" == "developer" ]; then
+            elif [ "$RHEL_SUB_METHOD" == "login" ]; then
                 subscription-manager register
             fi
-
             subscription-manager refresh
             subscription-manager attach --auto
-            echo "Checking subscription status..."
-            subscription_status=`subscription-manager status | grep 'Overall Status:' | awk '{ print $3 }'`
-
-            if [ "$subscription_status" == "Current" ]; then
-                break;
-            else
-                echo "Error subscription appears to be invalid please try again..."
-                exit 3
-            fi
-        done;
-
     fi
 }
 
@@ -125,6 +117,7 @@ function rhel_reposync() {
     done
     createrepo $repo_path
     echo "RHEL sync complete"
+    echo " "
 }
 
 function create_repo_config() {
@@ -138,14 +131,35 @@ cat <<EOF > /etc/httpd/conf.d/repo.conf
   </Directory>
 </VirtualHost>
 EOF
+    if [ -f /etc/httpd/conf.d/userdir.conf ]; then
     mv /etc/httpd/conf.d/userdir.conf /etc/httpd/conf.d/userdir.con.bak
+    fi
+    if [ -f /etc/httpd/conf.d/welcome.conf ]; then
     mv /etc/httpd/conf.d/welcome.conf /etc/httpd/conf.d/welcome.con.bak
+    fi
+    echo "Restarting httpd"
+    run_cmd systemctl restart httpd
+    run_cmd systemctl enable httpd
+    run_cmd systemctl start httpd
+    echo "Enabling httpd on firewall"
+    run_cmd firewall-cmd --permanent --add-service=http
+    echo "Reloading firewall"
+    run_cmd firewall-cmd --reload
+    echo " "
+
+}
+
+function cleaning() {
+    echo "Cleaning dnf"
+    dnf clean all
+    echo " "
 }
 
 root_check
 
 if [ -z "$RHEL_ACTIVATIONKEY" ] && [ -z "$RHEL_ORGANIZATION" ]; then
     subscription_prompts
+    attach_repos
 else
     register
     attach_repos
@@ -155,3 +169,4 @@ install_deps
 install_epel
 rhel_reposync
 create_repo_config
+cleaning
